add rate model to sdk
