+++ packages/sdk/src/across/index.ts
@@ -1,4 +1,5 @@
 export * as feeCalculator from "./feeCalculator";
+export * as rateModel from "./rateModel";
 export { default as LpFeeCalculator } from "./lpFeeCalculator";
 export * as gasFeeCalculator from "./gasFeeCalculator";
 export * as utils from "./utils";
